* time measurements in network code is now done with unsigned integral types (reduce amount of type conversions and floating point divisions)
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@4559 37977431-3df6-0310-b722-df95706aa16b

+++ rts/System/Net/UDPConnection.h
@@ -68,8 +68,8 @@ public:
 private:
 	void Init();
 	
-	float lastSendTime;
-	float lastReceiveTime;
+	unsigned lastSendTime;
+	unsigned lastReceiveTime;
 	
 	typedef boost::ptr_map<int,RawPacket> packetMap;
 	/// all packets with number <= nextAck arrived at the other end
@@ -92,7 +92,7 @@ private:
 	packetMap waitingPackets;
 	int lastInOrder;
 	int lastNak;
-	float lastNakTime;
+	unsigned lastNakTime;
 	std::queue<RawPacket*> msgQueue;
 
 	/** Our socket.
